requirements.txt
# Core vector database and ML dependencies
qdrant-client==1.10.1
sentence-transformers==3.0.1
numpy>=1.24.0,<2.0.0

# Data processing
pandas>=2.0.0,<3.0.0
openpyxl>=3.1.0

# Configuration and utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Optional: Additional ML and NLP libraries for enhanced functionality
# torch>=2.0.0  # PyTorch backend for sentence-transformers
# transformers>=4.30.0  # Hugging Face transformers library
# scikit-learn>=1.3.0  # For additional ML utilities
# plotly>=5.15.0  # For visualization (optional)
# streamlit>=1.28.0  # For web UI (optional)
